% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_asserted_distributions.R
\name{tw_asserted_distributions}
\alias{tw_asserted_distributions}
\alias{tw_ad}
\title{Asserted distributions}
\usage{
tw_asserted_distributions(
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_ad(csv = FALSE, token = NULL, project_token = NULL, page = 0, per = 50, ...)
}
\arguments{
\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}

\item{asserted_distribution_id}{(integer, vector) filter by asserted distribution id}

\item{descendants}{(boolean) include descendants}

\item{geo_json}{(boolean) filter by geojson}

\item{geographic_area_id}{(integer, vector) filter by geographic area id}

\item{geographic_area_mode}{(boolean)}

\item{otu_id}{(integer, vector) filter by otu id}

\item{presence}{(boolean) filter by presence}

\item{radius}{(integer) filter by meters around shape boundary}

\item{taxon_name_id}{(integer, vector) filter by taxon name id}

\item{wkt}{(string) filter by well known text}

\item{citations}{(boolean) filter by asserted distributions with citations}

\item{citation_documents}{(boolean) filter by asserted distributions with citation documents}

\item{origin_citation}{(boolean) filter by asserted distributions with an origin citation}

\item{data_attribute_exact_value}{(boolean) require exact match on data attribute value}

\item{data_attribute_predicate_id}{(integer, vector) filter by data attribute predicate id}

\item{data_attribute_value}{(string, vector) filter by data attribute value}

\item{data_attributes}{(boolean) filter by asserted distributions with data attributes}

\item{identifier}{(string) filter by identifier}

\item{identifier_end}{(string) filter by identifier end}

\item{identifier_exact}{(boolean) filter by exact match on identifier}

\item{identifier_start}{(string) filter by identifier start}

\item{identifier_type}{(string, vector) filter by identifier class type (e.g., "Identifier::Local::CatalogNumber")}

\item{identifiers}{(boolean) filter by asserted distributions with identifiers}

\item{local_identifiers}{(boolean) filter by asserted distributions with local identifiers}

\item{match_identifiers_delimiter}{(string) the delimiter used to separate identifiers (e.g., \n for line break, \t for tab)}

\item{match_identifiers_type}{(string) filter by internal or external identifier type}

\item{namespace_id}{(integer) filter by namespace id}

\item{keyword_id_and}{(integer, vector) filter by keyword id with and operator}

\item{keyword_id_or}{(integer, vector) filter by keyword id with or operator}

\item{tags}{(boolean) filter by asserted distributions with tags}
}
\value{
list
}
\description{
Asserted distributions

Asserted distributions
}
\examples{
\dontrun{
tw_asserted_distributions()
}
}
