% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_citations.R
\name{tw_citations}
\alias{tw_citations}
\title{Citations}
\usage{
tw_citations(
  citation_id = NULL,
  citation_object_id = NULL,
  citation_object_type = NULL,
  is_original = NULL,
  model_id = NULL,
  polymorphic = NULL,
  source_id = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{citation_id}{(integer, vector) filter by citation ID}

\item{citation_object_id}{(integer, vector) filter by citation object ID}

\item{citation_object_type}{(string, vector) filter by citation object type (AssertedDistribution, BiologicalAssociation, NomenclatureStatus, Otu, TaxonName, TaxonNameRelationship, TypeMaterial)}

\item{is_original}{(boolean) filter by original citation}

\item{model_id}{(string) filter by model ID}

\item{polymorphic}{(boolean) filter by polymorphic citation}

\item{source_id}{(integer) filter by source ID}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Citations
}
\examples{
\dontrun{
tw_citations()
}
}
