% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_collecting_events.R
\name{tw_collecting_events}
\alias{tw_collecting_events}
\alias{tw_ce}
\title{Collecting events}
\usage{
tw_collecting_events(
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_ce(csv = FALSE, token = NULL, project_token = NULL, page = 0, per = 50, ...)

tw_collecting_events(
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_ce(csv = FALSE, token = NULL, project_token = NULL, page = 0, per = 50, ...)
}
\arguments{
\item{token}{(NULL)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}

\item{citation_documents}{(boolean) filter by collecting events with citation documents}

\item{citations}{(boolean) filter by collecting events with citations}

\item{collection_objects}{(boolean) filter by collecting events with collection objects}

\item{collector_id}{(integer, vector) filter by collector id}

\item{collector_id_or}{(integer, vector) filter by collector id (OR)}

\item{collectors}{(boolean) filter by collecting events with collectors}

\item{data_attribute_exact_value}{(string) filter by data attribute exact value}

\item{data_attribute_predicate_id}{(integer, vector) filter by data attribute predicate id}

\item{data_attribute_value}{(string, vector) filter by data attribute value}

\item{data_attributes}{(boolean) filter by collecting events with data attributes}

\item{document_label}{(string)}

\item{elevation_precision}{(numeric)}

\item{end_date}{(string) filter by end date}

\item{end_date_day}{(integer)}

\item{end_date_month}{(integer)}

\item{end_date_year}{(integer)}

\item{field_notes}{(string)}

\item{formation}{(string)}

\item{geographic_area}{(boolean) filter by collecting events with geographic area}

\item{geographic_area_id}{(integer, vector) filter by geographic area id}

\item{geographic_area_mode}{(boolean)}

\item{georeferences}{(boolean) filter by collecting events with georeferences}

\item{group}{(string)}

\item{identifier}{(string) filter by cached identifier}

\item{identifier_end}{(string) filter by identifier end}

\item{identifier_exact}{(boolean) filter by exact match on identifier}

\item{identifier_start}{(string) filter by identifier start}

\item{identifier_type}{(string, vector) filter by identifier type}

\item{identifiers}{(boolean) filter by collecting events with identifiers}

\item{image_id}{(integer, vector) filter by image id}

\item{images}{(boolean) filter by collecting events with images}

\item{in_labels}{(string) wildcard wrapped search against any label}

\item{in_verbatim_locality}{(string) wildcard wrapped search against verbatim locality}

\item{keyword_id_and}{(integer, vector) filter by keyword id (AND)}

\item{keyword_id_or}{(integer, vector) filter by keyword id (OR)}

\item{lithology}{(string)}

\item{local_identifiers}{(boolean) filter by collecting events with local identifiers}

\item{match_identifiers_delimiter}{(string) the delimiter for the match_identifiers parameter}

\item{match_identifiers_type}{(string) the type of match to perform on the identifiers (internal or identifier)}

\item{maximum_elevation}{(numeric) filter by maximum elevation in meters}

\item{md5_of_verbatim_label}{(string)}

\item{md5_verbatim_label}{(string) filter by md5 hash of verbatim label}

\item{member}{(string)}

\item{meta_prioritize_geographic_area}{(string)}

\item{min_ma}{(numeric)}

\item{minimum_elevation}{(numeric) filter by minimum elevation in meters}

\item{namespace_id}{(integer) filter by namespace id}

\item{note_exact}{(boolean) filter by exact match on note}

\item{note_text}{(string) filter by note text}

\item{notes}{(boolean) filter by collecting events with notes}

\item{origin_citation}{(boolean) filter by collecting events with origin citations}

\item{otu_id}{(integer, vector) filter by otu id}

\item{print_label}{(string)}

\item{radius}{(integer) filter by radius in meters}

\item{recent}{(boolean) filter by recent collecting events}

\item{start_date}{(string) filter by start date}

\item{start_date_day}{(integer)}

\item{start_date_month}{(integer)}

\item{start_date_year}{(integer)}

\item{tags}{(boolean) filter by collecting events with tags}

\item{time_end_hour}{(integer)}

\item{time_end_minute}{(integer)}

\item{time_end_second}{(integer)}

\item{time_start_hour}{(integer)}

\item{time_start_minute}{(integer)}

\item{time_start_second}{(integer)}

\item{use_max}{(boolean)}

\item{use_min}{(boolean)}

\item{verbatim_collectors}{(string)}

\item{verbatim_datum}{(string)}

\item{verbatim_date}{(string)}

\item{verbatim_elevation}{(string)}

\item{verbatim_geolocation_uncertainty}{(numeric)}

\item{verbatim_habitat}{(string)}

\item{verbatim_label}{(string)}

\item{verbatim_latitude}{(numeric)}

\item{verbatim_locality}{(string)}

\item{verbatim_longitude}{(numeric)}

\item{verbatim_method}{(string)}

\item{verbatim_trip_identifier}{(string)}

\item{wkt}{(string) filter by well-known text describing the search area}
}
\value{
list

list
}
\description{
Collecting events

Collection events

Collection events
}
\examples{
\dontrun{
tw_collecting_events()
}
\dontrun{
tw_collecting_events()
}
}
