% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_contents.R
\name{tw_contents}
\alias{tw_contents}
\title{Contents}
\usage{
tw_contents(
  citations = NULL,
  citation_documents = NULL,
  origin_citation = NULL,
  exact = NULL,
  image_id = NULL,
  images = NULL,
  otu_id = NULL,
  text = NULL,
  topic_id = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{citations}{(boolean) filter by contents that have citations}

\item{citation_documents}{(boolean) filter by contents that have citation documents}

\item{origin_citation}{(boolean) filter by contents that have an origin citation}

\item{exact}{(boolean) require exact match on search term}

\item{image_id}{(integer, vector) filter by image id}

\item{images}{(boolean) filter by contents that have images}

\item{otu_id}{(integer, vector) filter by otu id}

\item{text}{(string) filter by text}

\item{topic_id}{(integer, vector) filter by topic id}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Contents
}
\examples{
\dontrun{
tw_contents()
}
}
