% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_data_attributes.R
\name{tw_data_attributes}
\alias{tw_data_attributes}
\alias{tw_da}
\title{Data attributes}
\usage{
tw_data_attributes(
  attribute_subject_id = NULL,
  attribute_subject_type = NULL,
  controlled_vocabulary_term_id = NULL,
  data_attribute_id = NULL,
  type = NULL,
  import_predicate = NULL,
  value = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_da(
  attribute_subject_id = NULL,
  attribute_subject_type = NULL,
  controlled_vocabulary_term_id = NULL,
  data_attribute_id = NULL,
  type = NULL,
  import_predicate = NULL,
  value = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{attribute_subject_id}{(integer) filter by attribute subject ID}

\item{attribute_subject_type}{(string) filter by attribute subject type (e.g., TaxonName)}

\item{controlled_vocabulary_term_id}{(integer, vector) filter by controlled vocabulary term ID}

\item{data_attribute_id}{(integer, vector) filter by data attribute ID}

\item{type}{(string) filter by data attribute type}

\item{import_predicate}{(string) filter by import predicate}

\item{value}{(string) filter by data attribute value}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Data attributes

Data attributes
}
\examples{
\dontrun{
tw_data_attributes()
}
}
