% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_depictions.R
\name{tw_depictions}
\alias{tw_depictions}
\title{Depictions}
\usage{
tw_depictions(
  depiction_id = NULL,
  depiction_object_id = NULL,
  depiction_object_type = NULL,
  image_id = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{depiction_id}{(integer, vector) filter by depiction id}

\item{depiction_object_id}{(integer, vector) filter by depiction object id}

\item{depiction_object_type}{(string, vector) filter by depiction object type}

\item{image_id}{(integer, vector) filter by image id}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Depictions
}
\examples{
\dontrun{
tw_depictions()
}
}
