% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_downloads.R
\name{tw_downloads}
\alias{tw_downloads}
\alias{tw_download}
\title{Downloads}
\usage{
tw_downloads(
  download_id = NULL,
  download_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_download(
  download_id = NULL,
  download_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{download_id}{(integer, vector) filter by download id}

\item{download_type}{(string, vector) filter by download type}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Downloads

Downloads
}
\examples{
\dontrun{
tw_downloads()
}
}
