% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_dwc_occurrences.R
\name{tw_dwc_occurrences}
\alias{tw_dwc_occurrences}
\alias{tw_dwc}
\title{Darwin Core occurrences}
\usage{
tw_dwc_occurrences(
  dwc_occurrence_id = NULL,
  dwc_occurrence_object_id = NULL,
  dwc_occurrence_object_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_dwc(
  dwc_occurrence_id = NULL,
  dwc_occurrence_object_id = NULL,
  dwc_occurrence_object_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{dwc_occurrence_id}{(integer, vector) filter by Darwin Core occurrence ID}

\item{dwc_occurrence_object_id}{(integer, vector) filter by Darwin Core occurrence object ID}

\item{dwc_occurrence_object_type}{(string, vector) filter by Darwin Core occurrence object type}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Darwin Core occurrences

Darwin Core occurrences
}
\examples{
\dontrun{
tw_people(last_name = "Smith")
}
}
