% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{tw_GET}
\alias{tw_GET}
\title{Perform a GET request to the TaxonWorks API}
\usage{
tw_GET(
  url,
  path = NULL,
  query = list(),
  headers = list(),
  csv = FALSE,
  vector_params = c(),
  ...
)
}
\arguments{
\item{url}{the base URL}

\item{path}{the endpoint path}

\item{query}{a list of query parameters}

\item{headers}{a list of headers}
}
\value{
a tibble of results
}
\description{
Perform a GET request to the TaxonWorks API
}
\keyword{internal}
