% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_identifiers.R
\name{tw_identifiers}
\alias{tw_identifiers}
\title{Identifiers}
\usage{
tw_identifiers(
  identifier = NULL,
  identifier_id = NULL,
  identifier_object_id = NULL,
  identifier_object_type = NULL,
  namespace_id = NULL,
  namespace_name = NULL,
  namespace_short_name = NULL,
  object_global_id = NULL,
  query_string = NULL,
  type = NULL,
  polymorphic = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{identifier}{(string, vector) filter by one or more identifiers}

\item{identifier_id}{(integer, vector) filter by one or more identifier IDs (e.g., the internal ID from the identifiers table in TaxonWorks)}

\item{identifier_object_id}{(integer, vector) filter by one or more object IDs (e.g., the internal ID from any related table in TaxonWorks)}

\item{identifier_object_type}{(string, vector) filter by one or more object types (e.g., Otu, CollectionObject, TaxonName, etc.)}

\item{namespace_id}{(integer, vector) filter by one or more namespace IDs}

\item{namespace_name}{(string) filter by a namespace name}

\item{namespace_short_name}{(string) filter by a namespace short name}

\item{object_global_id}{(string) filter by an object global ID}

\item{query_string}{(string) filter by a query string on the cached identifier (wildcards allowed)}

\item{type}{(string) filter by identifier type (e.g., Identifier::Local::CatalogNumber)}

\item{polymorphic}{(boolean) filter by polymorphic identifiers}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Identifiers
}
\examples{
\dontrun{
tw_identifiers(identifier_id=c(1375290,1375299,1375300))
}
}
