% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_images.R
\name{tw_images}
\alias{tw_images}
\title{Images}
\usage{
tw_images(
  biocuration_class_id = NULL,
  collection_object_id = NULL,
  collection_object_scope = NULL,
  depictions = NULL,
  depiction_object_type = NULL,
  image_id = NULL,
  otu_id = NULL,
  citations = NULL,
  citation_documents = NULL,
  origin_citation = NULL,
  keyword_id_and = NULL,
  keyword_id_or = NULL,
  tags = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{biocuration_class_id}{(integer, vector) filter by one or more biocuration class IDs}

\item{collection_object_id}{(integer, vector) filter by one or more collection object IDs}

\item{collection_object_scope}{(string) filter by collection object scope (e.g., CollectionObject, Observations, CollectingEvent)}

\item{depictions}{(boolean) filter by images that are depictions}

\item{depiction_object_type}{(string) filter by depiction object type}

\item{image_id}{(integer, vector) filter by one or more image IDs}

\item{otu_id}{(integer, vector) filter by one or more OTU IDs}

\item{citations}{(boolean) filter by images that have citations}

\item{citation_documents}{(boolean) filter by images that have citation documents}

\item{origin_citation}{(boolean) filter by images that have origin citations}

\item{keyword_id_and}{(integer, vector) filter by one or more keyword IDs (AND)}

\item{keyword_id_or}{(integer, vector) filter by one or more keyword IDs (OR)}

\item{tags}{(boolean) filter by images that have tags}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Images
}
\examples{
\dontrun{
tw_images(identifier_id=c(1375290,1375299,1375300))
}
}
