% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R, R/tw_notes.R
\name{tw_notes}
\alias{tw_notes}
\alias{tw_note}
\title{Title}
\usage{
tw_notes(
  note_id = NULL,
  note_object_id = NULL,
  note_object_type = NULL,
  text = NULL,
  polymorphic = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_note(
  note_id = NULL,
  note_object_id = NULL,
  note_object_type = NULL,
  text = NULL,
  polymorphic = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_notes(
  note_id = NULL,
  note_object_id = NULL,
  note_object_type = NULL,
  text = NULL,
  polymorphic = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_note(
  note_id = NULL,
  note_object_id = NULL,
  note_object_type = NULL,
  text = NULL,
  polymorphic = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{note_id}{(integer, vector) filter by note id}

\item{note_object_id}{(integer, vector) filter by note object id}

\item{note_object_type}{(string, vector) filter by note object type}

\item{text}{(string) filter by note text}

\item{polymorphic}{(boolean) filter by polymorphic notes}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list

list
}
\description{
Title

Title

Notes

Notes
}
\examples{
\dontrun{
tw_people(last_name="Smith")
}
\dontrun{
tw_notes(last_name="Smith")
}
}
