% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_observations.R
\name{tw_observations}
\alias{tw_observations}
\alias{tw_obs}
\title{Observations}
\usage{
tw_observations(
  character_state_id = NULL,
  collection_object_id = NULL,
  descendants = NULL,
  descriptor_id = NULL,
  observation_id = NULL,
  observation_matrix_id = NULL,
  observation_object_global_id = NULL,
  observation_object_type = NULL,
  observation_type = NULL,
  otu_id = NULL,
  taxon_name_id = NULL,
  citations = NULL,
  citation_documents = NULL,
  data_depictions = NULL,
  depictions = NULL,
  identifiers = NULL,
  notes = NULL,
  note_text = NULL,
  note_exact = NULL,
  protocols = NULL,
  tags = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_obs(
  character_state_id = NULL,
  collection_object_id = NULL,
  descendants = NULL,
  descriptor_id = NULL,
  observation_id = NULL,
  observation_matrix_id = NULL,
  observation_object_global_id = NULL,
  observation_object_type = NULL,
  observation_type = NULL,
  otu_id = NULL,
  taxon_name_id = NULL,
  citations = NULL,
  citation_documents = NULL,
  data_depictions = NULL,
  depictions = NULL,
  identifiers = NULL,
  notes = NULL,
  note_text = NULL,
  note_exact = NULL,
  protocols = NULL,
  tags = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{character_state_id}{(integer) filter by character state ID}

\item{collection_object_id}{(integer, vector) filter by collection object ID}

\item{descendants}{(boolean) filter by descendants}

\item{descriptor_id}{(integer, vector) filter by descriptor ID}

\item{observation_id}{(integer, vector) filter by observation ID}

\item{observation_matrix_id}{(integer, vector) filter by observation matrix ID}

\item{observation_object_global_id}{(string) filter by observation object global ID}

\item{observation_object_type}{(string, vector) filter by observation object type (Extract, CollectionObject, Otu)}

\item{observation_type}{(string, vector) filter by observation type (Observation::Qualitative, Observation::Presence, Observation::Continuous, Observation::Sample, Observation::Media, Observation::FreeText)}

\item{otu_id}{(integer, vector) filter by OTU ID}

\item{taxon_name_id}{(integer, vector) filter by taxon name ID}

\item{citations}{(boolean) filter by observations that have citations}

\item{citation_documents}{(boolean) filter by observations that have citation documents}

\item{data_depictions}{(boolean) filter by observations that have data depictions}

\item{depictions}{(boolean) filter by observations that have depictions}

\item{identifiers}{(boolean) filter by observations that have identifiers}

\item{notes}{(boolean) filter by observations that have notes}

\item{note_text}{(string) filter by note text}

\item{note_exact}{(boolean) filter by exact match on note}

\item{protocols}{(boolean) filter by observations that have protocols}

\item{tags}{(boolean) filter by observations that have tags}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Observations

Observations
}
\examples{
\dontrun{
tw_observations()
}
}
