% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_otus.R
\name{tw_otus}
\alias{tw_otus}
\alias{tw_otu}
\title{Otus}
\usage{
tw_otus(
  id = NULL,
  subresource = NULL,
  asserted_distribution_ids = NULL,
  biological_association_ids = NULL,
  citations = NULL,
  citations_documents = NULL,
  data_attribute_exact_value = NULL,
  data_attribute_predicate_id = NULL,
  data_attribute_value = NULL,
  data_attributes = NULL,
  data_attributes_attributes = NULL,
  image_id = NULL,
  images = NULL,
  keyword_id_and = NULL,
  keyword_id_or = NULL,
  name = NULL,
  note_exact = NULL,
  note_text = NULL,
  notes = NULL,
  origin_citations = NULL,
  otu_id = NULL,
  tags = NULL,
  taxon_name_classification_id = NULL,
  taxon_name_id = NULL,
  taxon_name_relationship_id = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_otu(
  id = NULL,
  subresource = NULL,
  asserted_distribution_ids = NULL,
  biological_association_ids = NULL,
  citations = NULL,
  citations_documents = NULL,
  data_attribute_exact_value = NULL,
  data_attribute_predicate_id = NULL,
  data_attribute_value = NULL,
  data_attributes = NULL,
  data_attributes_attributes = NULL,
  image_id = NULL,
  images = NULL,
  keyword_id_and = NULL,
  keyword_id_or = NULL,
  name = NULL,
  note_exact = NULL,
  note_text = NULL,
  notes = NULL,
  origin_citations = NULL,
  otu_id = NULL,
  tags = NULL,
  taxon_name_classification_id = NULL,
  taxon_name_id = NULL,
  taxon_name_relationship_id = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{id}{(integer) access an OTU by ID}

\item{subresource}{(string) access an API subresource (e.g., inventory/dwc)}

\item{asserted_distribution_ids}{(string) filter by asserted distribution ids}

\item{biological_association_ids}{(string) filter by biological association ids}

\item{citations}{(boolean) filter by otus with citations}

\item{citations_documents}{(boolean) filter by otus with citation documents}

\item{data_attribute_exact_value}{(boolean) require exact match on data attribute value}

\item{data_attribute_predicate_id}{(integer) filter by data attribute predicate id}

\item{data_attribute_value}{(string) filter by data attribute value}

\item{data_attributes}{(boolean) filter by otus with data attributes}

\item{data_attributes_attributes}{(string) filter by data attributes attributes}

\item{image_id}{(integer) filter by image id}

\item{images}{(boolean) filter by otus with images}

\item{keyword_id_and}{(integer) filter by keyword id with and operator}

\item{keyword_id_or}{(integer) filter by keyword id with or operator}

\item{name}{(string) filter by name}

\item{note_exact}{(boolean) require exact match on note text}

\item{note_text}{(string) filter by note text}

\item{notes}{(boolean) filter by otus with notes}

\item{origin_citations}{(boolean) filter by otus with origin citations}

\item{otu_id}{(integer) filter by otu id}

\item{tags}{(boolean) filter by otus with tags}

\item{taxon_name_classification_id}{(string) filter by taxon name classification ids}

\item{taxon_name_id}{(integer) filter by taxon name id}

\item{taxon_name_relationship_id}{(string) filter by taxon name relationship ids}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Otus

Otus
}
\examples{
\dontrun{
tw_otus()
}
}
