% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_people.R
\name{tw_people}
\alias{tw_people}
\title{People}
\usage{
tw_people(
  active_after_year = NULL,
  active_before_year = NULL,
  born_after_year = NULL,
  born_before_year = NULL,
  died_after_year = NULL,
  died_before_year = NULL,
  first_name = NULL,
  identifier = NULL,
  identifier_end = NULL,
  identifier_exact = NULL,
  identifier_start = NULL,
  last_name = NULL,
  last_name_starts_with = NULL,
  person_wildcard = NULL,
  role = NULL,
  project_id = NULL,
  name = NULL,
  tags = NULL,
  user_date_end = NULL,
  user_date_start = NULL,
  data_attribute_exact_value = NULL,
  data_attribute_predicate_id = NULL,
  data_attribute_value = NULL,
  data_attributes = NULL,
  note_exact = NULL,
  note_text = NULL,
  notes = NULL,
  keyword_id_and = NULL,
  keyword_id_or = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{active_after_year}{(integer) filter by active after year}

\item{active_before_year}{(integer) filter by active before year}

\item{born_after_year}{(integer) filter by born after year}

\item{born_before_year}{(integer) filter by born before year}

\item{died_after_year}{(integer) filter by died after year}

\item{died_before_year}{(integer) filter by died before year}

\item{first_name}{(string) filter by first name}

\item{identifier}{(string) filter by identifier}

\item{identifier_end}{(string) filter by identifier end}

\item{identifier_exact}{(boolean) filter by exact match on identifier}

\item{identifier_start}{(string) filter by identifier start}

\item{last_name}{(string) filter by last name}

\item{last_name_starts_with}{(string) filter by last name starts with}

\item{person_wildcard}{(vector) filter by person wildcard}

\item{role}{(vector, string) filter by role}

\item{project_id}{(integer) filter by project id}

\item{name}{(string) filter by name}

\item{tags}{(boolean) filter by people with tags}

\item{user_date_end}{(string) filter by user date end}

\item{user_date_start}{(string) filter by user date start}

\item{data_attribute_exact_value}{(boolean) require exact match on data attribute value}

\item{data_attribute_predicate_id}{(integer, vector) filter by data attribute predicate id}

\item{data_attribute_value}{(string, vector) filter by data attribute value}

\item{data_attributes}{(boolean) filter by people with data attributes}

\item{note_exact}{(boolean) require exact match on note text}

\item{note_text}{(string) filter by note text}

\item{notes}{(boolean) filter by people with notes}

\item{keyword_id_and}{(integer, vector) filter by keyword id with and operator}

\item{keyword_id_or}{(integer, vector) filter by keyword id with or operator}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
People
}
\examples{
\dontrun{
tw_people(last_name="Smith")
}
}
