% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_projects.R
\name{tw_projects}
\alias{tw_projects}
\title{Projects}
\usage{
tw_projects()
}
\value{
list with two slots
\itemize{
\item \code{result} (data.frame/tibble): results, a zero row data.frame
if no results found
\item \code{meta} (data.frame/tibble): number of results found
}
}
\description{
Projects
}
\examples{
tw_projects()
}
