% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_tags.R
\name{tw_tags}
\alias{tw_tags}
\alias{tw_tag}
\title{Tags}
\usage{
tw_tags(
  tag_id = NULL,
  tag_object_id = NULL,
  tag_object_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_tag(
  tag_id = NULL,
  tag_object_id = NULL,
  tag_object_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{tag_id}{(integer, vector) filter by tag id}

\item{tag_object_id}{(integer, vector) filter by tag object id}

\item{tag_object_type}{(string, vector) filter by tag object type}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Tags

Tags
}
\examples{
\dontrun{
tw_tags()
}
}
