% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_taxon_name_relationships.R
\name{tw_taxon_name_relationships}
\alias{tw_taxon_name_relationships}
\alias{tw_tnr}
\title{Taxon Name Relationships}
\usage{
tw_taxon_name_relationships(
  object_taxon_name_id = NULL,
  subject_taxon_name_id = NULL,
  taxon_name_id = NULL,
  taxon_name_relationship_set = NULL,
  taxon_name_relationship_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_tnr(
  object_taxon_name_id = NULL,
  subject_taxon_name_id = NULL,
  taxon_name_id = NULL,
  taxon_name_relationship_set = NULL,
  taxon_name_relationship_type = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{object_taxon_name_id}{(integer, vector) filter by object taxon name id}

\item{subject_taxon_name_id}{(integer, vector) filter by subject taxon name id}

\item{taxon_name_id}{(integer, vector) filter by taxon name id}

\item{taxon_name_relationship_set}{(string) filter by taxon name relationship set ('validating', 'invalidating', 'exceptions')}

\item{taxon_name_relationship_type}{(string, vector) filter by taxon name relationship type (e.g., TaxonNameRelationship::Iczn::Invalidating)}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Taxon Name Relationships

Taxon Name Relationships
}
\examples{
\dontrun{
tw_taxon_name_relationships()
}
}
