% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_taxon_names.R
\name{tw_taxon_names_autocomplete}
\alias{tw_taxon_names_autocomplete}
\alias{tw_tn_auto}
\title{Taxon Names Autocomplete}
\usage{
tw_taxon_names_autocomplete(
  term = NULL,
  exact = NULL,
  no_leaves = NULL,
  nomenclature_group = NULL,
  parent_id = NULL,
  project_id = NULL,
  type = NULL,
  valid = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)

tw_tn_auto(
  term = NULL,
  exact = NULL,
  no_leaves = NULL,
  nomenclature_group = NULL,
  parent_id = NULL,
  project_id = NULL,
  type = NULL,
  valid = NULL,
  csv = FALSE,
  token = NULL,
  project_token = NULL,
  page = 0,
  per = 50,
  ...
)
}
\arguments{
\item{term}{(string) filter by search term}

\item{exact}{(boolean) filter by exact match on name}

\item{no_leaves}{(boolean) filter by names having no descendants}

\item{nomenclature_group}{(string, vector) filter by nomenclature group (e.g., Family, Genus, Species)}

\item{parent_id}{(integer, vector) filter by parent id}

\item{project_id}{(integer) filter by project id}

\item{type}{(string, vector) filter by type (e.g., Protonym, Combination)}

\item{valid}{(boolean) filter by valid or invalid names}

\item{token}{(string) a user token (set with the TW_USER_TOKEN environment variable)}

\item{project_token}{(string) a project token (set with the TW_PROJECT_TOKEN environment variable)}

\item{page}{(integer) requested number of offset records (Default: 0)}

\item{per}{(integer) requested number of maximum records to be returned (Default: 50, Maximum: 10000)}

\item{...}{curl options passed on to \code{\link[httr2]{verb-GET}}}
}
\value{
list
}
\description{
Taxon Names Autocomplete

Taxon Names Autocomplete
}
\examples{
\dontrun{
tw_taxon_names(name="Lycorma delicatula", valid=TRUE)
}
}
